functions-python/reverse_geolocation_populate/requirements.txt
# Common packages
functions-framework==3.*
google-cloud-logging
psycopg2-binary==2.9.6
aiohttp~=3.10.5
asyncio~=3.4.3
urllib3~=2.2.2
requests~=2.32.3
attrs~=23.1.0
pluggy~=1.3.0
certifi~=2024.7.4

# SQL Alchemy and Geo Alchemy
SQLAlchemy==2.0.23
geoalchemy2==0.14.7

# Google specific packages for this function
google-cloud-pubsub
google-cloud-datastore
google-cloud-storage
google-cloud-bigquery
cloudevents~=1.10.1

# Additional packages for this function
gtfs-kit
pandas
pycountry